Society-Management-App/requirements.txt
APScheduler==3.8.0
asgiref==3.4.1
black==21.9b0
click==8.0.1
Django==3.2.7
django-cors-headers==3.10.0
django-environ==0.8.1
django-rest-framework==0.1.0
djangorestframework==3.12.4
mypy-extensions==0.4.3
mysqlclient==2.0.3
pathspec==0.9.0
Pillow==8.3.2
platformdirs==2.4.0
python-dateutil==2.8.2
pytz==2021.1
regex==2021.9.30
six==1.16.0
sqlparse==0.4.2
tomli==1.2.1
typing-extensions==3.10.0.2
tzlocal==2.1
